+++ src/main/java/fll/web/schedule/ProcessScheduleConstraints.java
@@ -7,20 +7,28 @@
 package fll.web.schedule;
 
 import java.io.IOException;
+import java.sql.Connection;
+import java.sql.SQLException;
 import java.util.Set;
 
+import javax.sql.DataSource;
+
+import fll.Tournament;
+import fll.db.TournamentParameters;
+import fll.util.FLLInternalException;
+import fll.web.ApplicationAttributes;
+import fll.web.AuthenticationContext;
+import fll.web.BaseFLLServlet;
+import fll.web.SessionAttributes;
+import fll.web.UserRole;
+import fll.web.WebUtils;
 import jakarta.servlet.ServletContext;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-
-import fll.web.AuthenticationContext;
-import fll.web.BaseFLLServlet;
-import fll.web.SessionAttributes;
-import fll.web.UserRole;
-import fll.web.WebUtils;
+import jakarta.servlet.jsp.PageContext;
 
 /**
  * Process the results of scheduleConstraints.jsp.
@@ -28,6 +36,18 @@ import fll.web.WebUtils;
 @WebServlet("/schedule/ProcessScheduleConstraints")
 public class ProcessScheduleConstraints extends BaseFLLServlet {
 
+  public static void populateContext(final ServletContext application,
+                                     final PageContext page) {
+    final DataSource datasource = ApplicationAttributes.getDataSource(application);
+    try (Connection connection = datasource.getConnection()) {
+      final Tournament tournament = Tournament.getCurrentTournament(connection);
+      final int numSeedingRounds = TournamentParameters.getNumSeedingRounds(connection, tournament.getTournamentID());
+      page.setAttribute("numSeedingRounds", numSeedingRounds);
+    } catch (final SQLException e) {
+      throw new FLLInternalException("Error reading from the database", e);
+    }
+  }
+  
   @Override
   protected void processRequest(final HttpServletRequest request,
                                 final HttpServletResponse response,
@@ -43,22 +63,20 @@ public class ProcessScheduleConstraints extends BaseFLLServlet {
     final UploadScheduleData uploadScheduleData = SessionAttributes.getNonNullAttribute(session, UploadScheduleData.KEY,
                                                                                         UploadScheduleData.class);
     try {
-
-      final String changeTimeDurationStr = WebUtils.getNonNullRequestParameter(request, "changeTimeDuration");
-      final int changeTimeDuration = Integer.parseInt(changeTimeDurationStr);
+      final int changeTimeDuration = WebUtils.getIntRequestParameter(request, "changeTimeDuration");
       uploadScheduleData.getSchedParams().setChangetimeMinutes(changeTimeDuration);
 
-      final String performanceChangeTimeDurationStr = WebUtils.getNonNullRequestParameter(request,
-                                                                                          "performanceChangeTimeDuration");
-      final int performanceChangeTimeDuration = Integer.parseInt(performanceChangeTimeDurationStr);
+      final int performanceChangeTimeDuration = WebUtils.getIntRequestParameter(request,
+                                                                                "performanceChangeTimeDuration");
       uploadScheduleData.getSchedParams().setPerformanceChangetimeMinutes(performanceChangeTimeDuration);
 
-      final String performanceDurationStr = WebUtils.getNonNullRequestParameter(request, "performanceDuration");
-      final int performanceDuration = Integer.parseInt(performanceDurationStr);
+      final int performanceDuration = WebUtils.getIntRequestParameter(request, "performanceDuration");
       uploadScheduleData.getSchedParams().setPerformanceMinutes(performanceDuration);
 
-      WebUtils.sendRedirect(application, response, "chooseScheduleHeaders.jsp");
+      final int numPerformanceRuns = WebUtils.getIntRequestParameter(request, "numPerformanceRuns");
+      uploadScheduleData.setNumPerformanceRuns(numPerformanceRuns);
 
+      WebUtils.sendRedirect(application, response, "chooseScheduleHeaders.jsp");
     } finally {
       session.setAttribute(UploadScheduleData.KEY, uploadScheduleData);
     }